requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
requests
opencv-python==4.8.1.78
numpy==1.24.3
groq==0.4.2 